refactor: controller 구축, 인터페이스화 완료, 각 controller 마다 show 하는 멤버변수 생성
