+++ src/main/java/subway/view/MainScreen.java
@@ -1,7 +1,6 @@
 package subway.view;
 
 import subway.Constants;
-import subway.Load;
 import subway.domain.Line;
 import subway.domain.LineRepository;
 import subway.domain.Station;
@@ -19,7 +18,15 @@ public class MainScreen implements Screen {
         return instance;
     }
 
-    private void printTransitMap() {
+    @Override
+    public String show() {
+        System.out.println(Constants.MAIN_SCREEN_USER_PROMPT);
+        String userInput = UserInputNumberSelection.createUserSelectionInput(
+                Constants.COUNT_MAIN_USER_PROMPT);
+        return userInput;
+    }
+
+    public void printTransitMap() {
         System.out.println("\n## 지하철 노선도");
         for (Line line : LineRepository.lines()) {
             System.out.println("[INFO] " + line.getName() + "\n[INFO] ---");
@@ -28,26 +35,5 @@ public class MainScreen implements Screen {
             }
             System.out.println();
         }
-    }
-
-    @Override
-    public void start() {
-        System.out.println(Constants.MAIN_SCREEN_USER_PROMPT);
-        int userInput = UserInputNumberSelection.createUserSelectionInput(
-                Constants.COUNT_MAIN_USER_PROMPT, Constants.QUIT);
-        if (userInput == Constants.USER_ANSWER_STATION_MANAGEMENT_SCREEN) {
-            Load.loadStationManagementScreen();
-        }
-        if (userInput == Constants.USER_ANSWER_LINE_MANAGEMENT_SCREEN) {
-            Load.loadLineManagementScreen();
-        }
-        if (userInput == Constants.USER_ANSWER_SECTION_MANAGEMENT_SCREEN) {
-            Load.loadSectionManagementScreen();
-        }
-        if (userInput == Constants.USER_ANSWER_PRINT_TRANSIT_MAP) {
-            printTransitMap();
-            start();
-        }
-        return;
     }
 }
