+++ src/main/java/subway/view/UserInputNumberSelection.java
@@ -12,12 +12,12 @@ public class UserInputNumberSelection {
         return scanner.nextLine();
     }
 
-    public static Integer createUserSelectionInput(int endInclusive, String returnOption) {
+    public static String createUserSelectionInput(int endInclusive) {
         boolean isValidInput = false;
         while (!isValidInput) {
             System.out.println("## 원하는 기능을 선택하세요.");
             try {
-                return validateUserSelectionInput(getUserInput(), endInclusive, returnOption);
+                return validateUserSelectionInput(getUserInput(), endInclusive);
             } catch (IllegalArgumentException e) {
                 System.err.println("[ERROR] 선택할 수 없는 기능입니다.");
                 isValidInput = false;
@@ -26,19 +26,14 @@ public class UserInputNumberSelection {
         return null;
     }
 
-    public static Integer validateUserSelectionInput(String userInput, int endInclusive, String returnOption) {
+    public static String validateUserSelectionInput(String userInput, int endInclusive) {
         userInput = userInput.toUpperCase();
-        if (userInput.equals(returnOption) && returnOption.equals(Constants.QUIT)) {
-            System.exit(0);
-        }
-        if (userInput.equals(returnOption) && returnOption.equals(Constants.BACK)) {
-            Load.loadMainScreen();
-            return null;
-        }
-        if (!userInput.matches( Constants.USER_INPUT_REGEX_START +
-                endInclusive + Constants.USER_INPUT_REGEX_END)) {
+
+        if (!(userInput.matches( Constants.USER_INPUT_REGEX_START +
+                endInclusive + Constants.USER_INPUT_REGEX_END))
+                || userInput.equals("B") || userInput.equals("Q")) {
             throw new IllegalArgumentException();
         }
-        return Integer.parseInt(userInput);
+        return userInput;
     }
 }
